profileコンポーネントリファクタリング。
