+++ frontend/src/features/profile/components/atoms/OtpAuthButton.tsx
@@ -1,9 +1,9 @@
 import { FC, memo, useState } from 'react';
-import { Box, Button, HStack, Text } from '@chakra-ui/react';
+import { Button } from '@chakra-ui/react';
 import { useOtpAuthCreate } from 'hooks/api';
 import { useIsOtpAuthEnabled } from 'hooks/api/auth/useIsOtpAuthEnabled';
 import { useOtpAuthDelete } from 'hooks/api/auth/useOtpAuthDelete';
-import { ModalOtpQrcode } from './ModalOtpQrcode';
+import { OtpQrcodeModal } from '../organisms/OtpQrcodeModal';
 
 export const OtpAuthButton: FC = memo(() => {
   const { isOtpAuthEnabled } = useIsOtpAuthEnabled();
@@ -26,30 +26,22 @@ export const OtpAuthButton: FC = memo(() => {
   };
 
   return (
-    <Box m={1}>
-      <HStack justify="center" align="center">
-        <Text>OTP Auth</Text>
-        {isOtpAuthEnabled ? (
-          <Button
-            size="xs"
-            fontSize="xs"
-            bg="green.200"
-            onClick={onClickInactivate}
-          >
-            active
-          </Button>
-        ) : (
-          <Button
-            size="xs"
-            fontSize="xs"
-            bg="red.200"
-            onClick={onClickActivate}
-          >
-            inactive
-          </Button>
-        )}
-      </HStack>
-      <ModalOtpQrcode isOpen={showFlag} onCloseModal={onCloseModal} />
-    </Box>
+    <>
+      {isOtpAuthEnabled ? (
+        <Button
+          size="xs"
+          fontSize="xs"
+          bg="green.200"
+          onClick={onClickInactivate}
+        >
+          active
+        </Button>
+      ) : (
+        <Button size="xs" fontSize="xs" bg="red.200" onClick={onClickActivate}>
+          inactive
+        </Button>
+      )}
+      <OtpQrcodeModal isOpen={showFlag} onCloseModal={onCloseModal} />
+    </>
   );
 });
