+++ frontend/src/features/profile/components/molecules/AvatarWithNickname.tsx
