+++ frontend/src/features/profile/components/molecules/GameResultCard.tsx
